==> 99_html_index.Rmd <==
---
title: 'Chimerism relapse results'
knit: (function(input_file, encoding) {
  out_dir <- 'docs';
  rmarkdown::render(input_file,
 encoding=encoding,
 output_file=file.path(dirname(input_file), out_dir, 'index.html'))})
author:
  - David C. Shyr^[Stanford Medicine, dcshyr@stanford.edu]
  - Simon Brewer^[University of Utah, simon.brewer@geog.utah.edu]
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    theme: flatly
    df_print: paged
    number_sections: true
    toc: false
    toc_float: false
---


- [Bone marrow chimerism plots][bmc_plots]: descriptive plots
- [Peripheral blood chimerism plots][pbc_plots]: descriptive plots
- [Random forest tuning][rf_tuning]: tuning and cross validation
- [Random forest results][rf_predict]: partial dependency plots
- [Hierarchical Random forest results][rf_merf]: Hierarchical/ME random forest
- [LIME results][rf_lime]: individual response plots (needs work)

[bmc_plots]: https://simonbrewer.github.io/chimerism_ml/plots_bmc.html
[pbc_plots]: https://simonbrewer.github.io/chimerism_ml/plots_pbc.html
[rf_tuning]: https://simonbrewer.github.io/chimerism_ml/rf_tuning.html
[rf_predict]: https://simonbrewer.github.io/chimerism_ml/rf_predict.html
[rf_merf]: https://simonbrewer.github.io/chimerism_ml/rf_merf.html
[rf_lime]: https://simonbrewer.github.io/chimerism_ml/rf_lime.html
